+++ mmaction/models/__init__.py
@@ -1,6 +1,7 @@
-from .backbones import (C3D, X3D, ResNet, ResNet2Plus1d, ResNet3d, ResNet3dCSN,
-                        ResNet3dLayer, ResNet3dSlowFast, ResNet3dSlowOnly,
-                        ResNetAudio, ResNetTIN, ResNetTSM)
+from .backbones import (C3D, X3D, MobileNetV2, MobileNetV2TSM, ResNet,
+                        ResNet2Plus1d, ResNet3d, ResNet3dCSN, ResNet3dLayer,
+                        ResNet3dSlowFast, ResNet3dSlowOnly, ResNetAudio,
+                        ResNetTIN, ResNetTSM)
 from .builder import (DETECTORS, build_backbone, build_detector, build_head,
                       build_localizer, build_loss, build_model, build_neck,
                       build_recognizer)
@@ -29,5 +30,5 @@ __all__ = [
     'TPN', 'TPNHead', 'build_loss', 'build_neck', 'AudioRecognizer',
     'AudioTSNHead', 'X3D', 'X3DHead', 'ResNet3dLayer', 'DETECTORS',
     'SingleRoIExtractor3D', 'BBoxHeadAVA', 'ResNetAudio', 'build_detector',
-    'ConvAudio', 'AVARoIHead'
+    'ConvAudio', 'AVARoIHead', 'MobileNetV2', 'MobileNetV2TSM'
 ]